+++ ChangeLog.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Voronoi neighborlist now includes periodic neighbors
 * Voronoi neighborlist computes weight according to the facet area in 3D
 * Box module exposes `getImage(vec)`
+* Voronoi module can compute and return cell volumes/areas
 
 ### Changed
 * Cluster module supports box argument in compute methods.